+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED] = 2024-02-14
+
+### Fixed
+
+- Unrecognized main stylesheet when using a [component-name].css name while having multiple stylesheets in a single
+  component
+
 ## [2.3.0] - 2024-02-01
 
 ### Added